chore: publish 0.54
commit 8b2da939e8c2077922fe39c94e5427193814449d
Author: Kevin <kevin@dendron.so>
Date:   Tue Aug 10 11:32:05 2021 -0700

    chore: update prep script

commit fce79fe797b75864c8f84126d5108b093cd58b2d
Author: Kevin <kevin@dendron.so>
Date:   Tue Aug 10 08:07:42 2021 -0700

    chore: publish minor

commit 14fcd58c4f3ac9f951b53e83ce6404724d745054
Author: Kevin <kevin@dendron.so>
Date:   Tue Aug 10 08:07:21 2021 -0700

    chore: bump 11ty

commit 29fe477fa1e1d79a1aa427ea600dbf57bbd8e5cc
Author: Kevin <kevin@dendron.so>
Date:   Tue Aug 10 08:07:08 2021 -0700

    chore: update build

commit d2e8c3a9793620edddde3e0ae1f3aca7bfeedc2e
Author: Kevin <kevin@dendron.so>
Date:   Tue Aug 10 07:57:38 2021 -0700

    chore: update no response bot

commit 87fac98afa505620b3d37c45925ac814871ceead
Author: Kevin <kevin@dendron.so>
Date:   Mon Aug 9 19:45:13 2021 -0700

    chore: add build scripts

commit 70453622ebddc3784fc25146f146bed452a0c33d
Author: Kevin <kevin@dendron.so>
Date:   Mon Aug 9 19:29:36 2021 -0700

    chore: add script to restore packages

commit 1b6672b9d55717b07fe4d8f99cc7362c5c33735b
Author: Kevin <kevin@dendron.so>
Date:   Mon Aug 9 19:20:18 2021 -0700

    chore: publish patch

commit 2d93e2a0adf4398cd3ddbf7a3fb10bd9b84acdde
Author: Kevin <kevin@dendron.so>
Date:   Mon Aug 9 19:20:02 2021 -0700

    chore: bump 11ty

commit ca81a6515942e8ef60a5df40f49d8dabc81121d5
Author: Kevin <kevin@dendron.so>
Date:   Mon Aug 9 19:18:58 2021 -0700

    chore: publish patch

commit cfe04d8dc4dca0f5ac0b1c11ec1ca6789bbfd290
Author: Kevin <kevin@dendron.so>
Date:   Mon Aug 9 19:18:43 2021 -0700

    chore: make nextjs private

commit 60e918636e91609cff84cd19f9396c31208967f9
Author: Kevin <kevin@dendron.so>
Date:   Mon Aug 9 19:16:37 2021 -0700

    chore: publish patch

commit d32cc4a66d309b34ea1d22fe175163c7497c1f06
Author: Kevin <kevin@dendron.so>
Date:   Mon Aug 9 19:16:17 2021 -0700

    chore: publish patch

commit 5edb09d6559ae6d68e1c6c31467c30af2407e9df
Author: Kevin <kevin@dendron.so>
Date:   Mon Aug 9 19:13:34 2021 -0700

    chore: publish patch

commit 74eace51944843c0b4277bbd2270e4b9b9ff83a6
Author: Kevin <kevin@dendron.so>
Date:   Mon Aug 9 19:08:32 2021 -0700

    chore: publish patch

commit d438be4f825b4e4cc8cccae24e21a666989274be
Author: Kevin <kevin@dendron.so>
Date:   Mon Aug 9 19:08:18 2021 -0700

    chore: bump 11ty

commit 87c92c78e470c45a13617edf26d606283dc1efca
Author: Kevin <kevin@dendron.so>
Date:   Mon Aug 9 18:44:40 2021 -0700

    chore: update yarn lock file

commit baa7d7eda498ba4715ef8dee311f57e1f277a05e
Author: Kevin <kevin@dendron.so>
Date:   Mon Aug 9 18:35:33 2021 -0700

    chore: publish patch

commit c236239beabd70c0c03633580b2c3198e431bdd9
Author: Kevin <kevin@dendron.so>
Date:   Mon Aug 9 18:35:16 2021 -0700

    chore: bump 11ty

+++ bootstrap/scripts/buildPatch.sh
@@ -1,10 +1,16 @@
+if [ $PUBLISH_ENDPOINT == "local" ]; then
+	echo "start verdaccio"
+	verdaccio &
+	FOO_PID=$!
+	echo "$FOO_PID"
+	sleep 3
+fi
 
-verdaccio &
-FOO_PID=$!
-echo "$FOO_PID"
+echo "building... upgrade: $UPGRADE_TYPE, endpoint: $PUBLISH_ENDPOINT"
 
-sleep 3
-LOG_LEVEL=info dendron dev build --upgradeType patch --publishEndpoint local
+LOG_LEVEL=info dendron dev build --upgradeType $UPGRADE_TYPE --publishEndpoint $PUBLISH_ENDPOINT
 
-echo "killing "
-kill $FOO_PID
+if [ $PUBLISH_ENDPOINT == "local" ]; then
+	echo "killing "
+	kill $FOO_PID
+fi
